Pushed all the logic into the exotic picker.
