+++ src/app/loadout-builder/filter/LockArmorAndPerks.tsx
@@ -20,6 +20,7 @@ import { isArmor2WithStats } from '../../loadout/item-utils';
 import { LoadoutBuilderAction } from '../loadout-builder-reducer';
 import LoadoutBucketDropTarget from '../LoadoutBucketDropTarget';
 import {
+  ItemsByBucket,
   LockableBuckets,
   LockedExclude,
   LockedExotic,
@@ -37,7 +38,7 @@ interface ProvidedProps {
   lockedMap: LockedMap;
   lockedMods: PluggableInventoryItemDefinition[];
   lockedExotic?: LockedExotic;
-  availableExotics?: DimItem[];
+  characterItems?: ItemsByBucket;
   unusableExotics?: DimItem[];
   lbDispatch: Dispatch<LoadoutBuilderAction>;
 }
@@ -69,7 +70,7 @@ function LockArmorAndPerks({
   lockedMods,
   buckets,
   stores,
-  availableExotics,
+  characterItems,
   unusableExotics,
   lockedExotic,
   isPhonePortrait,
@@ -271,7 +272,7 @@ function LockArmorAndPerks({
       {showExoticPicker &&
         ReactDom.createPortal(
           <ExoticPicker
-            availableExotics={availableExotics}
+            characterItems={characterItems}
             unusableExotics={unusableExotics}
             isPhonePortrait={isPhonePortrait}
             language={language}
